Map MNI152 coordinates to Colin27 coordinates via fsaverage.

Convenience function that maps MNI152 RAS coordinates to Colin27 RAS coordinates by chaining through fsaverage: MNI152 coords are first mapped to fsaverage vertices using mni152_coords_to_fsaverage, then the fsaverage vertices are mapped to Colin27 coordinates using fsaverage_vertices_to_colin27_coords.

Usage

mni152_coords_to_colin27_coords(
  coords,
  surface = "white",
  fs_home = Sys.getenv("FREESURFER_HOME"),
  silent = TRUE,
  simplify = FALSE
)

Arguments

coords

nx3 numeric matrix, the source RAS coordinates in the input image which must be in MNI152 space. The coords must be within the cortex, otherwise the mapping makes no sense and NaN values are returned for the respective coords.

surface

character string, the fsaverage surface (brain mesh) to load. Must be a valid FreeSurfer surface name like 'white', 'pial', 'orig, 'inflated'.

fs_home

character string, path to the FreeSurfer installation. Alternatively, a hemilist of freesurferformats::fs.surface instances like surface = list("lh"=mysurflh, "rh"=mysurfrh). Used to find the surfaces, at <fs_home>/subjects/fsaverage/surf/<hemi>.<surface>, where hemi is 'lh' and 'rh'. Can be NULL if 'surface' is a hemilist of fs.surface instances.

silent

logical, whether to suppress output messages in case of coords outside of cortex.

simplify

logical, whether to return a vector instead of a single-row matrix when only a single query coordinate is given.

Value

nx3 numeric matrix of Colin27 RAS coordinates. Also see the 'simplify' parameter.

Examples

if (FALSE) { # \dontrun{
  mni_ras = c(60.0, 0.0, 10.0);
  c27_coords = mni152_coords_to_colin27_coords(mni_ras, surface = "white");
} # }